refactor constants and update paths
