+++ miniprogram/common/constants.js
@@ -1,1 +1,7 @@
-module.exports = require('../../common/constants');
+module.exports = {
+  ADMIN_OPENIDS: ['admin_openid'],
+  LOG_STATUS: {
+    PENDING: 'pending',
+    APPROVED: 'approved',
+  },
+};
