+++ miniprogram/pages/common/constants.js
@@ -1,0 +1,1 @@
+module.exports = require('../../common/constants');
